Bug 5469: Install the table-miss flow entry into every OF1.3+ switch.
According to the OpenFlow 1.3 specification, OF1.3+ switch drops packets
unmatched by flow entries by default. So the VTN Manager needs to install
the table-miss flow entry into OF1.3+ switches so that unmatched packets
are punted to the controller.

+++ manager/implementation/src/main/java/org/opendaylight/vtn/manager/internal/flow/remove/RemovedFlowRemover.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 NEC Corporation.  All rights reserved.
+ * Copyright (c) 2015, 2016 NEC Corporation. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
@@ -7,6 +7,8 @@
  */
 
 package org.opendaylight.vtn.manager.internal.flow.remove;
+
+import static org.opendaylight.vtn.manager.internal.flow.remove.FlowRemoveContext.LOG;
 
 import com.google.common.base.Optional;
 
@@ -17,16 +19,21 @@ import org.opendaylight.vtn.manager.internal.TxContext;
 import org.opendaylight.vtn.manager.internal.util.DataStoreUtils;
 import org.opendaylight.vtn.manager.internal.util.IdentifiedData;
 import org.opendaylight.vtn.manager.internal.util.flow.FlowCache;
+import org.opendaylight.vtn.manager.internal.util.flow.FlowFinder;
 import org.opendaylight.vtn.manager.internal.util.flow.FlowUtils;
 import org.opendaylight.vtn.manager.internal.util.inventory.SalNode;
+import org.opendaylight.vtn.manager.internal.util.log.VTNLogLevel;
 
 import org.opendaylight.controller.md.sal.binding.api.ReadWriteTransaction;
 import org.opendaylight.controller.md.sal.common.api.data.LogicalDatastoreType;
 
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
 
+import org.opendaylight.yang.gen.v1.urn.opendaylight.vtn.flow.rev150410.VtnFlowId;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.vtn.impl.flow.rev150313.tenant.flow.info.VtnDataFlow;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.vtn.impl.flow.rev150313.vtn.flows.VtnFlowTable;
+
+import org.opendaylight.yang.gen.v1.urn.opendaylight.inventory.rev130819.nodes.Node;
 
 /**
  * An implementation of {@link FlowRemover} which removes a VTN data flow
@@ -34,9 +41,9 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.vtn.impl.flow.rev150313.vtn
  */
 public final class RemovedFlowRemover implements FlowRemover {
     /**
-     * The VTN data flow to be removed.
+     * The VTN flow ID that specifies the target VTN flow.
      */
-    private final IdentifiedData<VtnDataFlow>  dataFlow;
+    private final VtnFlowId  flowId;
 
     /**
      * The switch which removed the flow entry.
@@ -46,14 +53,13 @@ public final class RemovedFlowRemover implements FlowRemover {
     /**
      * Construct a new instance.
      *
-     * @param data   An {@link IdentifiedData} instance which contains the
-     *               VTN data flow to be removed.
+     * @param id     An {@link VtnFlowId} instance that specifies the VTN data
+     *               flow to be removed.
      * @param snode  A {@link SalNode} instance which specifies the switch
      *               which removed the flow entry.
      */
-    public RemovedFlowRemover(IdentifiedData<VtnDataFlow> data,
-                              SalNode snode) {
-        dataFlow = data;
+    public RemovedFlowRemover(VtnFlowId id, SalNode snode) {
+        flowId = id;
         flowNode = snode;
     }
 
@@ -73,22 +79,38 @@ public final class RemovedFlowRemover implements FlowRemover {
      */
     @Override
     public RemovedDataFlow removeDataFlow(TxContext ctx) throws VTNException {
-        // Check to see if the target data flow is present.
+        FlowCache fc;
+
+        // Ensure that the target node is present in opendaylight-inventory.
         ReadWriteTransaction tx = ctx.getReadWriteTransaction();
         LogicalDatastoreType oper = LogicalDatastoreType.OPERATIONAL;
-        InstanceIdentifier<VtnDataFlow> path = dataFlow.getIdentifier();
-        FlowCache fc;
-        Optional<VtnDataFlow> opt = DataStoreUtils.read(tx, oper, path);
+        Optional<Node> opt =
+            DataStoreUtils.read(tx, oper, flowNode.getNodeIdentifier());
         if (opt.isPresent()) {
-            // Delete the data flow.
-            fc = new FlowCache(opt.get());
-            tx.delete(oper, path);
+            // Search for the specified VTN data flow.
+            IdentifiedData<VtnDataFlow> data = new FlowFinder(tx).find(flowId);
+            if (data == null) {
+                ctx.log(LOG, VTNLogLevel.DEBUG,
+                        "Data flow not found: node={}, flow={}", flowNode,
+                        flowId.getValue());
+                fc = null;
+            } else {
+                // Delete the specified data flow.
+                fc = new FlowCache(data.getValue());
+                InstanceIdentifier<VtnDataFlow> path = data.getIdentifier();
+                tx.delete(oper, path);
 
-            // Clean up indices.
-            InstanceIdentifier<VtnFlowTable> tpath =
-                path.firstIdentifierOf(VtnFlowTable.class);
-            FlowUtils.removeIndex(tx, tpath, fc);
+                // Clean up indices.
+                InstanceIdentifier<VtnFlowTable> tpath =
+                    path.firstIdentifierOf(VtnFlowTable.class);
+                FlowUtils.removeIndex(tx, tpath, fc);
+            }
         } else {
+            // Nothing to do here.
+            // Rest of work will be done by VTN node REMOVED event handler.
+            ctx.log(LOG, VTNLogLevel.DEBUG,
+                    "Ignore FLOW_REMOVED on removed node: node={}, flow={}",
+                    flowNode, flowId.getValue());
             fc = null;
         }
 
@@ -100,10 +122,7 @@ public final class RemovedFlowRemover implements FlowRemover {
      */
     @Override
     public String getDescription() {
-        String tname = FlowUtils.getTenantName(dataFlow.getIdentifier());
-        return new StringBuilder("flow-removed[vtn=").
-            append(tname).append(", id=").
-            append(dataFlow.getValue().getFlowId().getValue()).
+        return new StringBuilder("flow-removed[id=").append(flowId.getValue()).
             append(", node=").append(flowNode).append(']').toString();
     }
 }
